refactor: change structure of the TaskModal component (taken out useEffect into castom hook, brought getOptions to utility)
