+++ client/src/utils/getOptions.ts
@@ -1,0 +1,6 @@
+export default function getOptions(array: string[]) {
+  return array.map((item) => ({
+    value: item,
+    label: item,
+  }));
+}
